// server/src/domain/enums/routes.enum.ts
export enum AuthRoutes {
  REGISTER = '/register',
  LOGIN = '/login',
  ADMIN_LOGIN = '/admin-login',
  GOOGLE_LOGIN = '/login/google',
  REFRESH = '/refresh',
  CHECK_AUTH = '/check-auth',
  FORGOT_PASSWORD = '/forgot-password',
  RESET_PASSWORD = '/reset-password',
  LOGOUT = '/logout',
  OTP_REQUEST = '/otp-request',
  OTP_VERIFY = '/otp-verify',
}